+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-- 0.1.0 (2024/12/21)
+- 0.2.0 (2025/6/5)
+
+  - Add a new button to "Download Markdown" for a mindmap.
+
+- 0.1.0 (2025/6/4)
 
   - Initial release
+  - Add a new button to "Expand all nodes" for a mindmap.